DataAnalyzer/requirements.txt
requests
kafka-python
psutil